quouta limit 회피 sleep method 삽입.
